feat(mocks): add support for node - restructure entrypoints
