+++ packages/mocks/setup.js
@@ -1,0 +1,2 @@
+export { mswRollupPlugin } from '@web/mocks/rollup-plugin.js';
+export { mockPlugin } from '@web/mocks/wds-plugin.js';
